--- Assets/Plugins/iOS/Editor/TealiumBuildPostProcess.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 46e8fd78eab1a4527918408830998951
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
